chore: extract theme vars to file for reuse
this prevents having duplicated themes colors everywhere and allows to change the default colors in a single, typed file

the definition file for the scss module vars is not generated automatically

scss needs to have the .module. in its name, otherwise importing it through vite returns a string instead of an object - leaving this here in case I do something similar in the future!