chore: fix build and update breaking change in prisma
Prisma 4 no longer provides the schema in the generated Prisma.dmmf, so we need to go fetch it ourselves

since @paljs/plugins is not yet updated to fix this issue, I fixed it myself by getting the proper dmmf myself

this change made the prisma selector function async, so I also updated its usage elsewhere in the code

testing it seems to work okay so far, even with an async subscription handler. neat!