chore: fix cors issues by using localstorage
this effectively prevents having cors issues at all

if no request token is sent to urql, it will not try to add auth to the operation, otherwise it will add it as an Authorization Header as `Bearer <token>`

few things had to be made for this to be smooth and working with both javascript enabled and disabled :
- detect if user has javascript by using temporary cookie, which gets deleted on every server fetch to prevent coming from enabled to disabled
- adding a new sessionToken store that uses the local storage store
- not using lucia's sameSite: none attribute as this won't be used anymore

essentially, the concept is that if javascript is enabled, then the functionnalities for realtime / client side fetching will work because otherwise they wouldn't anyway, so if javascript is enabled, use localstorage.

the cookies are still there to make it work for non-javascript folks. this makes it so I have to deal with both cookies and localstorage for auth management, but I'll make it work in my template then it should work everywhere