chore: route access check with route id and whitelist
this change makes it more strict with login : if the page exists and is not considered public by any of the frontend perm checkers, then do not access it.

doing this ensures that any user that isn't connected cannot access any page that isn't either whitelisted or can be navigated to using the on-screen navElements.

using the `route.id` also allows for 404 to pass through. Errors will however be swallowed : let's not make the app allow to error eh :)