refactor: do not use json datatype in forms anymore
this can now happen because the update to flowbite-svelte now includes a hidden `<select>` element, which works with formData.

while it is closer to work with javascript disabled, it still doesn't (maybe use the new `userHasJs` value? this would allow to make a if condition to show native select if JS disabled, MultiSelect otherwise)