feat: add api handling of `i18n`
a small quick refactor of the lucia factory user attributes means that now the data is more 1 to 1 with what is being produced, which allows the session to include database values rather than only the app defined DatabaseUserAttributes (this made some existing types breaking, so adding `| null` makes the errors go away)

small fix to the admin user edit page where the api errors were not showing up properly because the form was not updated

using a custom I18nModule to disable the built-in middleware... to add it back exactly as it is to prevent wrong middleware order

add i18n error management function by passing in a translator (works from both resolvers and services) which gets the current lang for you, and allows you to override it

the generated types of nestjs-i18n are placed next to the other generated types / data objects. fun!