feat!: move create / delete users as admin to tsrest
also added socket handling for created and deleted

important : changed `LuciaUser` type to use prisma's `User` type, omitting the id as Lucia adds its own. this adds the implication that creating users now requires to use all the database fields or mark them as null. the AuthService adds them all as null if not present, but its still good to know that it is now required and the returned `getUserAttributes` from `lucia.factory` is not a fake type, which could throw at runtime

the reason of the above change is because the user fetched is either a) all the fields from the users table OR b) the attributes provided when creating a user (it doesn't do a database roundtrip), so there was a few cases where a user could be missing fields, tripping up any user parsing, or missing data when the type system told us the data was there.

Another change was that `assertTsRestActionResultOK` now also has an optional `onNotOk` function that contains the information about what went wrong, the parsed error message, and the option to return your own fail data. it is useful to set custom form errors with a path that wouldn't have been there automatically, for example