feat: add email address changing mechanish
emails are used to confirm wheter or not the change is confirmed, as you need access to the new email address to change the email.

a new `textblock` handlebar helper was added to automatically format text with line breaks to use multiple `<p>` in the email formatter

tweaked `flash` getter to work in SSR, so flash is now also available for non-javascript users (there is still a fallback in case the method throws, which returns undefined, to see if that is still needed)

action results assertions now redirects when given an `event` in the arguments, as you should do with POST action requests. the plan is to eventually move `assertTsRestActionResultOK` to a local event function, so that you don't need to add the event every single time in the args for it to work