chore: add prisma-case-format auto handling for prisma
this tool allows to format the schema file using specified casings, so that the mapping can be automatically setup through different conventions

I love it as I want my database to be snake_case, while the code should be PascalCase and camelCase. this also handles the plurization of most fields, which is a time save for the auto prisma format.