feat: add automatic i18n page title
the title is set in the layout based on the current page route id, which is the sveltekit file system based route. using this, I can hook into the same system used for i18n key matching, and simply create a key that is based on the current route.

to set the page title, you simply have to create a i18n json file with the same route id of the current page, otherwise it will default to using the header without anything appended