fix(client): remove tailwind-purgecss
This is causing issues, I don't want to spend time trying to figure out which classes are being overly purged, and with their own doc :

This plugin will no longer be necessary after the release of Tailwind v4 as they are introducing a new Vite plugin that will detect and generate the tailwind classes based on the module graph! As such, this plugin will only support Tailwind v3.

therefore, removing this will only bloat the css because of flowbite until I get around to update tailwind to v4 (waiting on issue on flowbite-svelte)